Remove color to current task, link between calendar and task

+++ app/src/main/java/com/example/retime/CalendarTask.java
@@ -39,8 +39,8 @@ import java.util.Locale;
 public class CalendarTask extends AppCompatActivity {
 
     CalendarView calendar;
-    SQLiteDatabase db;
-    SQLiteOpenHelper openHelper;
+    SQLiteDatabase db, db2;
+    SQLiteOpenHelper openHelper, openHelper2;
     ImageButton task;
     Button savebtn, deletebtn;
     EditText firsttasket, firsttimeet, secondtasket, secondtimeet;
@@ -74,6 +74,8 @@ public class CalendarTask extends AppCompatActivity {
         secondtimeet = findViewById(R.id.secondtime);
         openHelper = new CalendarDatabase(this);
         db = openHelper.getWritableDatabase();
+        openHelper2 = new TaskDatabase(this);
+        db2 = openHelper2.getWritableDatabase();
         calendar = findViewById(R.id.calendarView);
         today = new SimpleDateFormat("dd/MM/yyyy").format(new Date(calendar.getDate()));
         todayarr = today.split("/");
@@ -89,7 +91,11 @@ public class CalendarTask extends AppCompatActivity {
             firsttimeet.setText(data.getString(2));
             secondtasket.setText(data.getString(3));
             secondtimeet.setText(data.getString(4));
-            notifimgr.notify(0, builder.build());
+            String currenttime = new SimpleDateFormat("kk:mm").format(Calendar.getInstance().getTime());
+            if (currenttime.substring(0,2).equals(data.getString(2)) || currenttime.substring(0,2).equals(data.getString(4))){
+                notifimgr.notify(1, builder.build());
+            }
+//            notifimgr.notify(0, builder.build());
         }
 
         calendar.setOnDateChangeListener(new CalendarView.OnDateChangeListener() {
@@ -131,6 +137,7 @@ public class CalendarTask extends AppCompatActivity {
         String firsttime = firsttimeet.getText().toString();
         String secondtask = secondtasket.getText().toString();
         String secondtime = secondtimeet.getText().toString();
+
         if (TextUtils.isEmpty(firsttask) && TextUtils.isEmpty(firsttime) && TextUtils.isEmpty(secondtask) && TextUtils.isEmpty(secondtime)){
             deletebtn.setVisibility(View.INVISIBLE);
         }
@@ -158,6 +165,7 @@ public class CalendarTask extends AppCompatActivity {
     }
 
     public void insertdata (String day, String firsttask, String firsttime, String secondtask, String secondtime) {
+
         ContentValues contentValues = new ContentValues();
         contentValues.put(CalendarDatabase.COL_1, day);
         contentValues.put(CalendarDatabase.COL_2, firsttask);
@@ -165,16 +173,93 @@ public class CalendarTask extends AppCompatActivity {
         contentValues.put(CalendarDatabase.COL_4, secondtask);
         contentValues.put(CalendarDatabase.COL_5, secondtime);
         long id = db.replace(CalendarDatabase.TABLE_NAME, null, contentValues);
+
+        if (!TextUtils.isEmpty(firsttime)) {
+            firsttime = Integer.toString(Integer.parseInt(firsttime) - 1);
+        }
+
+        if (!TextUtils.isEmpty(secondtime)){
+            secondtime = Integer.toString(Integer.parseInt(secondtime) - 1);
+        }
+
+        String ID1 = day.concat("/").concat(firsttime);
+        String ID2 = day.concat("/").concat(secondtime);
+
+        if (firsttime.equals(secondtime)){
+            ContentValues contentValues2 = new ContentValues();
+            contentValues2.put(TaskDatabase.COL_1, ID1);
+            contentValues2.put(TaskDatabase.COL_2, firsttask);
+            contentValues2.put(TaskDatabase.COL_3, "");
+            contentValues2.put(TaskDatabase.COL_4, "");
+            contentValues2.put(TaskDatabase.COL_5, secondtask);
+            contentValues2.put(TaskDatabase.COL_6, "");
+            contentValues2.put(TaskDatabase.COL_7, "");
+            contentValues2.put(TaskDatabase.COL_8, day);
+            contentValues2.put(TaskDatabase.COL_9, firsttime);
+            id = db2.replace(TaskDatabase.TABLE_NAME, null, contentValues2);
+        }
+        else {
+            ContentValues contentValues3 = new ContentValues();
+            ContentValues contentValues4 = new ContentValues();
+
+            contentValues3.put(TaskDatabase.COL_1, ID1);
+            contentValues3.put(TaskDatabase.COL_2, "");
+            contentValues3.put(TaskDatabase.COL_3, "");
+            contentValues3.put(TaskDatabase.COL_4, "");
+            contentValues3.put(TaskDatabase.COL_5, firsttask);
+            contentValues3.put(TaskDatabase.COL_6, "");
+            contentValues3.put(TaskDatabase.COL_7, "");
+            contentValues3.put(TaskDatabase.COL_8, day);
+            contentValues3.put(TaskDatabase.COL_9, firsttime);
+            if(!TextUtils.isEmpty(firsttask) && !TextUtils.isEmpty(firsttime)){
+                id = db2.replace(TaskDatabase.TABLE_NAME, null, contentValues3);
+            }
+
+            contentValues4.put(TaskDatabase.COL_1, ID2);
+            contentValues4.put(TaskDatabase.COL_2, "");
+            contentValues4.put(TaskDatabase.COL_3, "");
+            contentValues4.put(TaskDatabase.COL_4, "");
+            contentValues4.put(TaskDatabase.COL_5, secondtask);
+            contentValues4.put(TaskDatabase.COL_6, "");
+            contentValues4.put(TaskDatabase.COL_7, "");
+            contentValues4.put(TaskDatabase.COL_8, day);
+            contentValues4.put(TaskDatabase.COL_9, secondtime);
+            if(!TextUtils.isEmpty(secondtask) && !TextUtils.isEmpty(secondtime)){
+                id = db2.replace(TaskDatabase.TABLE_NAME, null, contentValues4);
+            }
+
+        }
+
         Toast.makeText(this, "save successfully", Toast.LENGTH_LONG).show();
 
     }
 
-    public void deletedata (String day) {
+    public void deletedata (String day, String firsttime, String secondtime) {
         long id = db.delete(CalendarDatabase.TABLE_NAME, CalendarDatabase.COL_1 + " = " + day, null);
         firsttasket.setText("");
         firsttimeet.setText("");
         secondtasket.setText("");
         secondtimeet.setText("");
+
+        if (!TextUtils.isEmpty(firsttime)) {
+            firsttime = Integer.toString(Integer.parseInt(firsttime) - 1);
+        }
+
+        if (!TextUtils.isEmpty(secondtime)){
+            secondtime = Integer.toString(Integer.parseInt(secondtime) - 1);
+        }
+
+        String ID1 = day.concat("/").concat(firsttime);
+        String ID2 = day.concat("/").concat(secondtime);
+
+        if (firsttime.equals(secondtime)){
+            id = db2.delete(TaskDatabase.TABLE_NAME, CalendarDatabase.COL_1 + " = '" + ID1 + "'", null);
+        }
+        else {
+            id = db2.delete(TaskDatabase.TABLE_NAME, CalendarDatabase.COL_1 + " = '" + ID1 + "'", null);
+            id = db2.delete(TaskDatabase.TABLE_NAME, CalendarDatabase.COL_1 + " = '" + ID2 + "'", null);
+        }
+
         Toast.makeText(this, "delete successfully", Toast.LENGTH_LONG).show();
     }
 
@@ -183,20 +268,43 @@ public class CalendarTask extends AppCompatActivity {
         @Override
         public void onClick(View v) {
             String firsttask = firsttasket.getText().toString();
-            String firsttime = firsttimeet.getText().toString();
-            String secondtask = secondtasket.getText().toString();
-            String secondtime = secondtimeet.getText().toString();
-            String todaydate = _dayofmonth.concat(_month).concat(_year);
-            if ((TextUtils.isEmpty(firsttask) && !TextUtils.isEmpty(firsttime)) || TextUtils.isEmpty(secondtask) && !TextUtils.isEmpty(secondtime)) {
-                Toast.makeText(CalendarTask.this, "Please enter task name", Toast.LENGTH_LONG).show();
-            } else {
+            String firsttime, secondtime;
+            if (!TextUtils.isEmpty(firsttimeet.getText().toString())){
+                firsttime = Integer.toString(Integer.parseInt(firsttimeet.getText().toString()));
+            }
+            else {
+                firsttime = "";
+            }
 
-                insertdata(todaydate, firsttask, firsttime, secondtask, secondtime);
-                if (TextUtils.isEmpty(firsttask) && TextUtils.isEmpty(firsttime) && TextUtils.isEmpty(secondtask) && TextUtils.isEmpty(secondtime)) {
-                    deletebtn.setVisibility(View.INVISIBLE);
-                    long id = db.delete(CalendarDatabase.TABLE_NAME, CalendarDatabase.COL_1 + " = " + todaydate, null);
-                } else {
-                    deletebtn.setVisibility(View.VISIBLE);
+            String secondtask = secondtasket.getText().toString();
+            if (!TextUtils.isEmpty(secondtimeet.getText().toString())){
+                secondtime = Integer.toString(Integer.parseInt(secondtimeet.getText().toString()));
+            }
+            else {
+                secondtime = "";
+            }
+
+            String todaydate = _dayofmonth.concat(_month).concat(_year);
+            if(TextUtils.isEmpty(firsttask) && !TextUtils.isEmpty(secondtask)){
+                Toast.makeText(CalendarTask.this, "Please fill in the first task", Toast.LENGTH_LONG).show();
+            }
+            else {
+                if ((TextUtils.isEmpty(firsttask) && !TextUtils.isEmpty(firsttime)) || TextUtils.isEmpty(secondtask) && !TextUtils.isEmpty(secondtime)) {
+                    Toast.makeText(CalendarTask.this, "Please enter task name", Toast.LENGTH_LONG).show();
+                }
+                else {
+                    if ((!TextUtils.isEmpty(firsttask) && (Integer.parseInt(firsttime) > 24 || Integer.parseInt(firsttime) < 1)) || (!TextUtils.isEmpty(secondtask) && (Integer.parseInt(secondtime) > 24 || Integer.parseInt(secondtime) < 1))){
+                        Toast.makeText(CalendarTask.this, "Please enter the time between 1 ~ 24", Toast.LENGTH_LONG).show();
+                    }
+                    else {
+                        insertdata(todaydate, firsttask, firsttime, secondtask, secondtime);
+                        if (TextUtils.isEmpty(firsttask) && TextUtils.isEmpty(firsttime) && TextUtils.isEmpty(secondtask) && TextUtils.isEmpty(secondtime)) {
+                            deletebtn.setVisibility(View.INVISIBLE);
+                            long id = db.delete(CalendarDatabase.TABLE_NAME, CalendarDatabase.COL_1 + " = " + todaydate, null);
+                        } else {
+                            deletebtn.setVisibility(View.VISIBLE);
+                        }
+                    }
                 }
             }
         }
@@ -207,7 +315,9 @@ public class CalendarTask extends AppCompatActivity {
         @Override
         public void onClick(View v) {
             String todaydate = _dayofmonth.concat(_month).concat(_year);
-            deletedata(todaydate);
+            String firsttime = firsttimeet.getText().toString();
+            String secondtime = secondtimeet.getText().toString();
+            deletedata(todaydate, firsttime, secondtime);
             deletebtn.setVisibility(View.INVISIBLE);
         }
     }
